Set setAllowOpen(false)
This is not a solution, of course, but a strong hint that
something is seriously wrong with the design of isThisType(): the
RandomAccessInputStream is never reused, and as a consequence, files
are read and re-read ad nauseam, just to find out what would be the
best reader.

Let the numbers speak: this brings the benchmark runtime from 317ms
down to 186ms.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>